refactor placement view logic into useViewed hook
